ArmPlatformPkg: Allocate VRAM as RuntimeServicesData
The UEFI specification allows the operating system (OS) to use the
Graphics Output Protocol (GOP) in the following scenarios:
 a. as part of the startup process and
 b. prior to loading of a high performance OS graphics driver

If the VRAM is allocated as BootServicesData, then it is unmapped on
exit boot services. This prevents GOP usage by the OS post exit boot
services (the second scenario); as it results in a crash when the VRAM
is accessed.

This patch fixes the issue by allocating VRAM as RuntimeServicesData.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Alexei Fedorov <alexei.fedorov@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Signed-off-by: Evan Lloyd <evan.lloyd@arm.com>