ArmPlatformPkg: Add HDLCD Support for Juno
The HDLCD Juno library implements the platform specific interfaces
for configuring the Juno HDLCD controller.

This patch is required to enable Graphics Output Protocol on Juno.

Note:
The HDLCD Juno Library does not configure the pixel clock and relies
on the default pixel clock (148.5 MHz) which is setup by the SCP to
support 1920 x 1080 resolution.

Known Issues:
Due to hardware timing interactions the display output is not stable
on certain monitors. This can cause a range of effects from occasional
glitches to screen blanking.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Alexei Fedorov <alexei.fedorov@arm.com>
Signed-off-by: Girish Pathak <girish.pathak@arm.com>
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Signed-off-by: Evan Lloyd <evan.lloyd@arm.com>