ArmPkg:Provide GetTriggerType/SetTriggerType functions
This change implements GetTriggerType and SetTriggerType functions
in ArmGicV2Dxe (GicV2GetTriggerType/GicV2SetTriggerType)
and ArmGicV3Dxe (GicV3GetTriggerType/GicV3SetTriggerType)

SetTriggerType configures the interrupt mode of an interrupt
as edge sensitive or level sensitive.

GetTriggerType function returns the mode of an interrupt.

The requirement for this change derives from a problem detected on ARM
Juno boards, but the change is of generic relevance.

NOTE: At this point the GICv3 code is not tested.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Girish Pathak <girish.pathak@arm.com>
Signed-off-by: Evan Lloyd <evan.lloyd@arm.com>
Tested-by: Girish Pathak <girish.pathak@arm.com>